+++ packages/chain-agnostic-permission/src/scope/constants.test.ts
@@ -54,6 +54,7 @@ describe('KnownRpcMethods', () => {
           "eth_uninstallFilter",
         ],
         "solana": Array [],
+        "tron": Array [],
       }
     `);
   });